fix:gpio改回

+++ luatos/components/luat/port/luat_gpio_idf5.c
@@ -99,7 +99,7 @@ int luat_gpio_setup(luat_gpio_t *gpio) {
     return 0;
 }
 
-int luat_gpio_set(int pin, uint8_t level) {
+int luat_gpio_set(int pin, int level) {
     if (pin == Luat_GPIO_MAX_ID)
         return 0;
     if (GPIO_CHECK(pin)) {
